# openapi.yaml
openapi: "3.1.0"
info:
  title: WSI Viewer Backend
  version: "1.0.0"
  description: |
    Whole-slide image viewer — tile streaming, session management,
    Tumour Content Analysis (TCA), per-slide annotations, and pathology review worklist.
servers:
  - url: http://localhost:8080
    description: Local dev
...
components:
  parameters:
    patient_id:
      name: patient_id
      in: query
      required: true
      schema: { type: string }
      description: Patient UUID
    event_id:
      name: event_id
      in: query
      required: true
      schema: { type: string }
      description: Encounter / event ID
    selected_slide_id:
      name: selected_slide_id
      in: query
      required: true
      schema: { type: string }
      description: Slide block ID
  schemas:
...
    SlideMetadata:
      type: object
      properties:
        width:           { type: integer }
        height:          { type: integer }
        mpp:             { type: number, nullable: true }
        objective_power: { type: integer, nullable: true }
        vendor:          { type: string, nullable: true }
    Slide:
      type: object
      properties:
        slide_id:  { type: string }
        block_id:  { type: string }
        filename:  { type: string, nullable: true }
        status:    { type: string }
        has_tca:   { type: boolean }
        meta_data: { $ref: '#/components/schemas/SlideMetadata', nullable: true }
    SlideList:
      type: object
      properties:
        slides:        { type: array, items: { $ref: '#/components/schemas/Slide' } }
        default_slide: { type: string, nullable: true }
    OverlayConfig:
      type: object
      properties:
        available:      { type: boolean }
        density_image:  { type: string, nullable: true }
        metadata:       { type: string, nullable: true }
        grid:           { type: string, nullable: true }
        slide_metadata: { $ref: '#/components/schemas/SlideMetadata', nullable: true }

    # ── Point (shared) ────────────────────────────────────────────────────
...
paths:
...
  /api/slides:
    get:
      tags: [wsi]
      summary: Get slide list for a session
      parameters:
        - $ref: '#/components/parameters/patient_id'
        - $ref: '#/components/parameters/event_id'
        - $ref: '#/components/parameters/selected_slide_id'
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: '#/components/schemas/SlideList' } } } }
        "502": { description: Pathology API unreachable }
        "503": { description: API credentials not configured }
  /api/thumbnail/{slide_name}:
    get:
      tags: [wsi]
      summary: Proxy thumbnail image from GCS
      parameters:
        - name: slide_name
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/patient_id'
        - $ref: '#/components/parameters/event_id'
        - $ref: '#/components/parameters/selected_slide_id'
      responses:
        "200": { description: PNG/JPEG thumbnail bytes }
        "404": { description: Slide or thumbnail not found }
  /api/raw_slides/{filename}:
    get:
      tags: [wsi]
      summary: Proxy Range request for tile streaming
      parameters:
        - name: filename
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/patient_id'
        - $ref: '#/components/parameters/event_id'
        - $ref: '#/components/parameters/selected_slide_id'
      responses:
        "200": { description: Full slide bytes }
        "206": { description: Partial content (Range response) }
        "404": { description: Slide not found in session }
    head:
      tags: [wsi]
      summary: HEAD — return Content-Length for tile planning
      parameters:
        - name: filename
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/patient_id'
        - $ref: '#/components/parameters/event_id'
        - $ref: '#/components/parameters/selected_slide_id'
      responses:
        "200": { description: Headers only }
  /api/slide-direct-url/{filename}:
    get:
      tags: [wsi]
      summary: Return signed GCS URL for direct browser access
      parameters:
        - name: filename
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/patient_id'
        - $ref: '#/components/parameters/event_id'
        - $ref: '#/components/parameters/selected_slide_id'
      responses:
        "200":
          description: Signed URL or null if unavailable
          content:
            application/json:
              schema:
                type: object
                properties:
                  url:       { type: string, nullable: true }
                  available: { type: boolean }
  /api/overlay-config/{slide_name}:
    get:
      tags: [wsi]
      summary: TCA overlay config (triggers asset download if needed)
      parameters:
        - name: slide_name
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/patient_id'
        - $ref: '#/components/parameters/event_id'
        - $ref: '#/components/parameters/selected_slide_id'
      responses:
        "200": { description: Overlay config, content: { application/json: { schema: { $ref: '#/components/schemas/OverlayConfig' } } } }
  /api/overlay-file/{filename}:
    get:
      tags: [wsi]
      summary: Serve extracted TCA overlay asset (PNG or JSON)
      parameters:
        - name: filename
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/patient_id'
        - $ref: '#/components/parameters/event_id'
        - $ref: '#/components/parameters/selected_slide_id'
      responses:
        "200": { description: File bytes }
        "404": { description: File not in overlay cache }

  # ── TCA ──────────────────────────────────────────────────────────────────
...